Split compose backend and speed up SOC-CA test keygen (#417)
Resolve the SonarCloud quality-gate findings on the deployment backend. docker_compose.py exceeded the 500-line module budget, so move the host/container query + inspect methods into ComposeQueryMixin (_compose_queries.py) and have DockerComposeBackend inherit it. Parameterize the bare dict / list[dict] annotations as dict[str, Any], collapse the four-return network/container inspect helpers to two returns via shared JSON decoders, and drop the redundant FileNotFoundError (a subclass of OSError) from the pull_images catch.

ensure_soc_certs minted a fresh 4096-bit RSA CA on every call and the suite drives that 27 times; that prime search is CPU-bound and stalls on a loaded host, which read as a hang. Make the CA and service key sizes a seam (production defaults unchanged at 4096 / 2048) and shrink them in tests via an autouse conftest fixture. test_soc_ca runs in ~1.6s.
